+++ public/admin/_app/immutable/nodes/24.Cfpqn5si.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/Bi29iOKp.js";export{e as component};